requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
tensorflow==2.16.1
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
jinja2==3.1.3
gunicorn==21.2.0 